Fix warnings about conversion from int to unsigned char/uint8_t
When compiling with cmake -DCMAKE_BUILD_TYPE=ASAN under Ubuntu 25.10
(gcc 15.2.0).

Explicitly cast these instances to uint8_t/unssigned char to silence this
warning.

Acked-by: Gert Doering <gert@greenie.muc.de>
Gerrit URL: https://gerrit.openvpn.net/c/openvpn/+/1328
Message-Id: <20251030193834.1186-1-gert@greenie.muc.de>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg34065.html

+++ src/openvpn/mroute.c
@@ -337,7 +337,7 @@ mroute_addr_mask_host_bits(struct mroute_addr *ma)
             }
             else
             {
-                ma->v6.addr.s6_addr[byte--] &= (0xFF << bits_to_clear);
+                ma->v6.addr.s6_addr[byte--] &= (uint8_t)(0xFF << bits_to_clear);
                 bits_to_clear = 0;
             }
         }
